% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dates.R
\name{add_dates}
\alias{add_dates}
\title{add_dates}
\usage{
add_dates(DF)
}
\arguments{
\item{DF}{input Data Frame or Data Table}
}
\description{
Add dates, day of year, water year, water year day to RHESSys output (as appropriate for time step)
}
\author{
Will Burke
}
