% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_runid_to_filename.R
\name{add_runID_to_filename}
\alias{add_runID_to_filename}
\title{Create a subfolder and add a runID to a filename}
\usage{
add_runID_to_filename(filename, runID = NULL)
}
\arguments{
\item{filename}{Path and name of file with extension}

\item{runID}{The unique ID to be appended to the filename.}
}
\description{
When conducting multiple runs, this function 1) creates the subfolder for
modified def and tec files, and 2) generates filenames with unique runID for
the def and tec files within each subfolder.
}
