% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_output_filter.R
\name{build_output_filter}
\alias{build_output_filter}
\title{build_output_filter}
\usage{
build_output_filter(
  timestep,
  output_format,
  output_path,
  output_filename,
  spatial_level,
  spatial_ID,
  variables
)
}
\arguments{
\item{timestep}{The timestep of the filter: 'daily', 'monthly', 'annual', 'hourly'}

\item{output_format}{The format for the RHESSys output files to be written: 'csv' or 'netcdf'}

\item{output_path}{The path where RHESSys output will be written.}

\item{output_filename}{The filename for RHESSys output}

\item{spatial_level}{The spatial level for output to be collected at/aggregated to: 'basin', 'hillslope', 'zone', 'patch', 'stratum'}

\item{spatial_ID}{IDs to subset the spatial units by.}

\item{variables}{The RHESSys internal variable names to output. If aggregating from a finer spatial level use syntax: '<spatial>.<varaible>'}
}
\description{
Creates a single output filter in R list format. This can be combined with other filters
and/or written to file to be read by RHESSys when running.
}
\author{
Will Burke
}
