% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_redefine.R
\name{build_redefine}
\alias{build_redefine}
\title{build_redefine}
\usage{
build_redefine(
  worldfile,
  out_file,
  vars = NULL,
  values = NULL,
  std_thin = NULL,
  patchID = NULL,
  strataID = NULL,
  veg_parm_ID = NULL,
  append = FALSE
)
}
\arguments{
\item{worldfile}{Source worldfile}

\item{out_file}{Destination file to write}

\item{vars}{variables to edit}

\item{values}{values to insert for variables}

\item{std_thin}{Value to insert for standard thinning state vars (replacement value or multiplier depending on usage)}

\item{patchID}{Patch ID(s) to apply redefine to, can be used alone or with other subsets}

\item{strataID}{Strata ID(s) to apply redefine to, can be used alone or with other subsets}

\item{veg_parm_ID}{veg parm ID to apply changes to, can be used alone or with other subsets}
}
\description{
Create a redefine worldfile
}
\author{
Will Burke
}
