% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_clim_base_file.R
\name{change_clim_base_file}
\alias{change_clim_base_file}
\title{Adds a dated sequence to the bottom of a clim base file}
\usage{
change_clim_base_file(clim_base_file, par_sets, file_name_ext = NULL)
}
\arguments{
\item{???}{???}

\item{???}{???}

\item{???}{???}
}
\description{
This function does ...
}

