% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_params}
\alias{check_params}
\title{check_params}
\usage{
check_params(rh_file, def_file)
}
\arguments{
\item{rh_file}{Path to the appropriate construct file in the RHESSys source code, e.g. construct_stratum_defaults.c}

\item{def_file}{Path to a appropriate parameter definition file to be compared to the RHESSys defaults}
}
\description{
Function to check if parameters are valid and compare them to default parameters from a version of RHESSys
}
\author{
Will Burke
}
