% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_rhessys.R
\name{cleanup_rhessys}
\alias{cleanup_rhessys}
\title{cleanup_rhessys}
\usage{
cleanup_rhessys(
  dir = NULL,
  output_name = NULL,
  copy = TRUE,
  destination = NULL
)
}
\arguments{
\item{dir}{Directory to look for output files within. To be safe function will not look recursively in subfolders.}

\item{output_name}{The base name of the rhessys output, used to select only one set of rhessy outputs
(still getting all of the varyious spatial + temporal levels for that output)}

\item{copy}{TRUE/FALSE should the output be copied or just deleted}

\item{destination}{Specify a destination for output to be copied to, if left NULL a folder will be
generated with a unique name with format rh_out_datetime}
}
\description{
Function to remove rhessys output, options to move it to a specific folder or one w unique ID,
only delete/move output based on a pattern
}
\author{
William Burke
}
