% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Code for evaluating model output}
\usage{
evaluation(sim_variables, obs_variables, obj_function_selection, input_folder,
  output_folder, ..., evaluation_type = c("single_objective_function",
  "composite", "pareto"))
}
\arguments{
\item{sim_variables}{????}

\item{obs_variables}{\code{obs} must have an equal number of values as \code{sim}.}
}
\description{
Description...
}
\details{
Future iterations will account for multiple parameters and/or multiple criteria.
Also, future iterations should will need to combine outputs from separate simulations (from parallelization)
into a single evaluated simulation.
}
