% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_option_sets.R
\name{generate_option_sets}
\alias{generate_option_sets}
\title{Produces data-frames for all parameter combinations and for each def file}
\usage{
generate_option_sets(parameter_method, input_hdr_list, input_rhessys,
  input_preexisting_table, input_def_list, input_standard_par_list,
  input_dated_seq_list)
}
\description{
Produces data-frames for all parameter combinations and for each def file
}
