% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_output_variables.R
\name{initialize_output_variables}
\alias{initialize_output_variables}
\title{Initialize output variables}
\usage{
initialize_output_variables(output_variables, output_folder)
}
\description{
This function nulls outputs from previous runs and initializes new output variables.
}
