% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_in_worldfile.R
\name{insert_in_worldfile}
\alias{insert_in_worldfile}
\title{Insert in Worldfile}
\usage{
insert_in_worldfile(
  world_in,
  world_out,
  insert,
  insert_loc,
  return_data = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{world_in}{Path to input worldfile. Cannot create from scratch (use RHESSysPreprocessing)}

\item{world_out}{Path to output worldfile. If set to FALSE, no file will be written.}

\item{insert}{Character vector of length = 2, containing state variable and value to be inserted into a worldfile.}

\item{insert_loc}{Location, in the form of an existing state variable name, for a new state variable(s) to be inserted following.}

\item{return_data}{TRUE/FALSE should an R data object be returned.}

\item{overwrite}{TRUE/FALSE should output worldfile overwrite existing file.}
}
\description{
Inserts new state variables and values into a worldfile at specified location(s).
}
\author{
Will Burke
}
