% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_clim.R
\name{IOin_clim}
\alias{IOin_clim}
\title{IOin_clim}
\usage{
IOin_clim(
  base_station_id = NULL,
  grid_cells = NULL,
  x_coordinate,
  y_coordinate,
  z_coordinate,
  effective_lai,
  screen_height,
  annual_prefix = "annual",
  num_non_critical_annual_sequences = 0,
  monthly_prefix = "monthly",
  num_non_critical_monthly_sequences = 0,
  daily_prefix = "daily",
  num_non_critical_daily_sequences = 0,
  hourly_prefix = "hourly",
  num_non_critical_hourly_sequences = 0
)
}
\arguments{
\item{base_station_id}{Base station ID. Used for station-based meteorological inputs.}

\item{grid_cells}{Number of grid cells. Used for gridded meteorological inputs.}

\item{x_coordinate}{X coordinate.}

\item{y_coordinate}{Y coordinate.}

\item{z_coordinate}{Z coordinate.}

\item{effective_lai}{Effective LAI.}

\item{screen_height}{Screen height.}

\item{annual_prefix}{Prefix for annual climate inputs.}

\item{num_non_critical_annual_sequences}{Number of non critical annual climate inputs. Defaults to 0.}

\item{monthly_prefix}{Prefix for monthly climate inputs.}

\item{num_non_critical_monthly_sequences}{Number of non critical annual climate inputs. Defaults to 0.}

\item{daily_prefix}{Prefix for daily climate inputs.}

\item{num_non_critical_daily_sequences}{Number of non critical annual climate inputs. Defaults to 0.}

\item{hourly_prefix}{Prefix for hourly climate inputs.}

\item{num_non_critical_hourly_sequences}{Number of non critical annual climate inputs. Defaults to 0.}
}
\description{
Generate input for run_rhessys climate basestation input
}
\author{
Will Burke
}
