% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_cmd_pars.R
\name{IOin_cmd_pars}
\alias{IOin_cmd_pars}
\title{IOin_cmd_pars}
\usage{
IOin_cmd_pars(
  m = 1,
  k = 1,
  soil_dep = 1,
  m_v = 1,
  k_v = 1,
  pa = 1,
  po = 1,
  gw1 = 1,
  gw2 = 1,
  vgseng1 = 1,
  vgseng2 = 1,
  vgseng3 = 1,
  n = 1,
  pct_range = 0.25
)
}
\arguments{
\item{m}{Decay of hydraulic conductivity with depth.}

\item{k}{Hydraulic conductivity at the surface.}

\item{soil_dep}{Soil depth.}

\item{m_v}{Multiplier to scale the vertical decay of hydraulic conductivity with depth.}

\item{k_v}{Multiplier to scale the vertical hydraulic conductivity at the surface.}

\item{pa}{Multiplier to scale the pore size index, set in soil def file.}

\item{po}{Multiplier to scale the psi air entry, set in soil def file.}

\item{gw1}{Multiplier on the sat_to_gw_coeff parameter set in the soil definition file
(representing the amount of water moving from the saturated store to the groundwater store).}

\item{gw2}{Multiplier on the gw_loss_coeff parameter in the hillslope default file
(representing the amount of water moving from the groundwater store to the stream).}

\item{vgseng1}{Multiples specific leaf areas.}

\item{vgseng2}{Multiplies the ratio of shaded to sunlit leaf area.}

\item{vgseng3}{Multiplier used only with the Dickenson algorithm of carbon allocation
(set with the epc.allocation_flag variable in the vegetation definition file). It changes
the allocation of net photosynthate sensitivity based on the current LAI. If not using the
Dickenson strategy of carbon allocation (i.e. using Waring or default Constant strategies),
set third value to 1.0. (i.e. -vgsen 1.0 2.0 1.0)}

\item{n}{The number of parameter sets to generate.}

\item{pct_range}{The percent range of variation from input values over which sampling (if any), will happen.}
}
\description{
Input standard command line soil parameters for RHESSys, see https://github.com/RHESSys/RHESSys/wiki/RHESSys-command-line-options
These parameters are MULTIPLIERS on the existing paramters in the definition files.
}
\author{
Will Burke
}
