% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_def_pars_latin_hypercube.R
\name{IOin_def_pars_latin_hypercube}
\alias{IOin_def_pars_latin_hypercube}
\title{IOin_def_pars_latin_hypercube}
\usage{
IOin_def_pars_latin_hypercube(..., rm_dup = TRUE)
}
\arguments{
\item{...}{Any number of lists, each containing 3 elements in format: list("<def file>", "<parameter name>", <c(n, min value, max value)>). The last element in the list is a vector containing the total number of parameter sets, minimum value of parameter sampling range, and maximum value of parameter sampling range. n must be equal across all lists.}

\item{rm_dup}{TRUE/FALSE should duplicate def file + variable entries be automatically removed? A warning will occur regardless.}
}
\description{
Generates multiple def file changes based on a sample of parameter values across the full range of each parameter.
}
\author{
Ryan Bart
}
