% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_def_pars_sobol.R
\name{IOin_def_pars_sobol}
\alias{IOin_def_pars_sobol}
\title{IOin_def_pars_sobol}
\usage{
IOin_def_pars_sobol(
  ...,
  nboot = 100,
  rm_dup,
  sobolfunction = sensitivity::sobol2007,
  return_sobol = FALSE
)
}
\arguments{
\item{...}{Any number of lists, each containing 3 elements in format: list("<def file>", "<parameter name>", <value>)}

\item{nboot}{The number of bootstraps to run for sobol2007}

\item{rm_dup}{TRUE/FALSE should duplicate def file + variable entries be automatically removed? A warning will occur regardless.}
}
\description{
Geneartes multiple def file changes based on sobol sensativity.
}
\author{
Will Burke
}
