% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_hdr.R
\name{IOin_hdr}
\alias{IOin_hdr}
\title{IOin_hdr}
\usage{
IOin_hdr(
  basin,
  hillslope,
  zone,
  soil,
  landuse,
  stratum,
  fire = NULL,
  basestations
)
}
\arguments{
\item{basin}{Path to basin parameter definition file.}

\item{hillslope}{Path to hillslope parameter definition file.(s)}

\item{zone}{Path to zone parameter definition file(s).}

\item{soil}{Path to soil parameter definition file(s).}

\item{landuse}{Path to landuse parameter definition file(s).}

\item{stratum}{Path to stratum parameter definition file(s).}

\item{fire}{Path to fire parameter definition file.}

\item{basestations}{Path to basin climate basestation file(s).}
}
\description{
Creates a header file based on specified parameter definition files and climate basestations.
}
\author{
Will Burke
}
