% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_output_filters.R
\name{IOin_output_filters}
\alias{IOin_output_filters}
\title{IOin_output_filters}
\usage{
IOin_output_filters(..., file_name = NULL, filter_in = NULL)
}
\arguments{
\item{...}{Any number of valid filter objects (R lists), to be written to the same output filter file.}

\item{file_name}{The name of the output filter file created. if left NULL an autogenerated file name will be used.}

\item{filter_in}{Path to an existing yaml filter file.}
}
\description{
Reads in and/or combines filters into a single R list filter object. Also checks that all filters are valid and contain the required fields
with valid options when possible to check.
}
\author{
Will Burke
}
