% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_output_vars.R
\name{IOin_output_vars}
\alias{IOin_output_vars}
\title{IOin_output_vars}
\usage{
IOin_output_vars(...)
}
\arguments{
\item{...}{Any number of two element character vectors or dataframes, each specifying the output file using "pd", "pdg", "cd", "cdg"
style format in the first item/column and the variable of interest in the second item/column.}
}
\description{
IOin_output_vars
}
\author{
Will Burke
}
