% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_rhessys_input.R
\name{IOin_rhessys_input}
\alias{IOin_rhessys_input}
\title{IOin_rhessys_input}
\usage{
IOin_rhessys_input(
  version,
  tec_file,
  world_file,
  world_hdr_prefix,
  flowtable,
  start,
  end,
  output_folder,
  output_prefix,
  commandline_options,
  world_hdr_path = NULL,
  prefix_command = NULL
)
}
\arguments{
\item{version}{Path to compiled RHESSys binary.}

\item{tec_file}{Path and name of input tec file. If also supplying a dataframe of tec events as created via IOin_tec_std() to run_rhessys_single(), a new directory will be created
at the supplied path and name, with a new tec file within it. If not supplying a dataframe of tec events, this input refers to an existing tec file to be used.}

\item{world_file}{Path and name of input world file.}

\item{world_hdr_prefix}{Path and name for new folder where newly created hdr files will go.}

\item{flowtable}{Path and name of input tec file.}

\item{start}{Start date of simulation.}

\item{end}{End date of simulation.}

\item{output_folder}{Path to folder where simulation output will go.}

\item{output_prefix}{Prefix for output files.}

\item{commandline_options}{Commandline options to be passed to RHESSys, e.x. '-g' or '-p'}

\item{world_hdr_path}{Path to where worldfile header file should be created/located. If left NULL,
will use world_hdr_prefix as both path and file name.}

\item{prefix_command}{A shell command to be run previous to the RHESSys command line call.
This can be used to source a shell script, which itself can run multiple commands if needed.}
}
\description{
Basic inputs to run RHESSys
}
\author{
Will Burke
}
