% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_std_pars.R
\name{IOin_std_pars}
\alias{IOin_std_pars}
\title{IOin_std_pars}
\usage{
IOin_std_pars(m, k, m_v, k_v, pa, po, gw1, gw2, n = 1, pct_range = 0.25)
}
\arguments{
\item{m}{Decay of hydraulic conductivity with depth}

\item{k}{Hydraulic conductivity at the surface}

\item{m_v}{Multiplier to scale the vertical decay of hydraulic conductivity with depth}

\item{k_v}{Multiplier to scale the vertical hydraulic conductivity at the surface}

\item{pa}{Multiplier to scale the pore size index, set in soil def file.}

\item{po}{Multiplier to scale the psi air entry, set in soil def file.}

\item{gw1}{Multiplier on the sat_to_gw_coeff parameter set in the soil definition file
(representing the amount of water moving from the saturated store to the groundwater store).}

\item{gw2}{Multiplier on the gw_loss_coeff parameter in the hillslope default file
(representing the amount of water moving from the groundwater store to the stream).}
}
\description{
Input standard soil parameters for RHESSys, see https://github.com/RHESSys/RHESSys/wiki/RHESSys-command-line-options
}
\author{
Will Burke
}
