% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_tec_all_options.R
\name{IOin_tec_all_options}
\alias{IOin_tec_all_options}
\title{IOin_tec_all_options}
\usage{
IOin_tec_all_options(
  print_hourly_on = NULL,
  print_hourly_growth_on = NULL,
  print_hourly_off = NULL,
  print_hourly_growth_off = NULL,
  print_daily_on = NULL,
  print_daily_growth_on = NULL,
  print_daily_off = NULL,
  print_daily_growth_off = NULL,
  print_monthly_on = NULL,
  print_monthly_off = NULL,
  print_yearly_on = NULL,
  print_yearly_growth_on = NULL,
  print_yearly_off = NULL,
  print_yearly_growth_off = NULL,
  redefine_strata = NULL,
  redefine_world = NULL,
  redefine_world_multiplier = NULL,
  redefine_world_thin_remain = NULL,
  redefine_world_thin_harvest = NULL,
  redefine_world_thin_snags = NULL,
  roads_on = NULL,
  roads_off = NULL,
  output_current_state = NULL,
  data = NULL
)
}
\arguments{
\item{print_hourly_on}{Start printing hourly output data}

\item{print_hourly_growth_on}{Start printing hourly carbon-related output data}

\item{print_hourly_off}{Stop printing hourly output data}

\item{print_hourly_growth_off}{Stop printing hourly carbon-related output data}

\item{print_daily_on}{Start printing daily output data}

\item{print_daily_growth_on}{Start printing daily carbon-related output data}

\item{print_daily_off}{Stop printing daily output data}

\item{print_daily_growth_off}{Stop printing daily carbon-related output data}

\item{print_monthly_on}{Start printing monthly output data}

\item{print_monthly_off}{Stop printing monthly output data}

\item{print_yearly_on}{Start printing yearly output data}

\item{print_yearly_growth_on}{Start printing yearly carbon-related output data}

\item{print_yearly_off}{Stop printing yearly output data}

\item{print_yearly_growth_off}{Stop printing yearly carbon-related output data}

\item{redefine_strata}{Redefine strata based on a file}

\item{redefine_world}{Redefine world based on a file}

\item{redefine_world_multiplier}{Redefine world based on a multiplier of
existing worldfile values}

\item{redefine_world_thin_remain}{Redefine world based on a thinning event
where removed vegetation is transferred to CWD and litter stores}

\item{redefine_world_thin_harvest}{Redefine world based on a thinning event
where removed vegetation exits watershed}

\item{redefine_world_thin_snags}{Redefine world based on a thinning event
where stems are left in place.}

\item{roads_on}{Start processing roads}

\item{roads_off}{Stop processing roads}

\item{output_current_state}{Output worldfile}

\item{data}{Optional dataframe containing tec arguments as column names and a
single row containing dates. The dataframe may contain list-columns with
vectors when more than one date is needed. All other arguments to the
function are treated as null when data argument is not null.}
}
\description{
Generic input function to construct a dataframe for all possible RHESSys
temporal event control (TEC) commands except csv commands. The function is
flexible, allowing the user to construct complex tec files. However, due to
the number of potential combinations of tec events, there are few checks on
appropriateness of tec event sequences (for example, turning off a tec event
before it is turned on). Full list of tec commands in RHESSys is available at
init/construct_tec.c
}
\section{Note}{
 All arguments accept a date or vector of dates in the format of
 "year month day hour" e.g. c("2000 10 1 1", "2001 9 30 1").

 For more information on redefine worlds, see wiki page
 \href{https://github.com/RHESSys/RHESSys/wiki/Redefining-the-worldfile}{Redefining-the-worldfile}
}

\author{
Ryan R Bart
}
