% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clim_base_file.R
\name{make_clim_base_file}
\alias{make_clim_base_file}
\title{Adds a dated sequence to the bottom of a clim base file}
\usage{
make_clim_base_file(input_clim_base, clim_base_path, input_dated_seq,
  clim_dated_ext)
}
\arguments{
\item{???}{???}

\item{???}{???}

\item{???}{???}
}
\description{
This function does ...
}
