% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dated_seq.R
\name{make_dated_seq}
\alias{make_dated_seq}
\title{Make Dated Sequence File from R}
\usage{
make_dated_seq(input_dated_seq, dated_seq_file)
}
\arguments{
\item{input_dated_seq}{Data frame ...}

\item{dated_seq_file}{File ...}
}
\description{
Description...
}
