% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dated_seq_file.R
\name{make_dated_seq_file}
\alias{make_dated_seq_file}
\title{Make Dated Sequence File from R}
\usage{
make_dated_seq_file(dated_seq_file, dated_seq_data)
}
\arguments{
\item{dated_seq_file}{Name and directory of dated sequence file to be created}

\item{dated_seq_data}{Data frame ...}
}
\description{
Description...
}

