% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hdr_file.R
\name{make_hdr_file}
\alias{make_hdr_file}
\title{Produces components of a hdr file}
\usage{
make_hdr_file(master_table, path_initial, num_files, default_file)
}
\arguments{
\item{master_table}{Data frame of a single row with references to def file number}

\item{path_initial}{Path to the non-differiated def file}

\item{default_file}{}
}
\description{
This function generates the lines in a hdr file associated with a single def file.
}

