% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_option_set_combinations.R
\name{make_option_set_combinations}
\alias{make_option_set_combinations}
\title{Generates a dataframe of parameter set combinations}
\usage{
make_option_set_combinations(input_list, parameter_method)
}
\arguments{
\item{input_list}{List of parameters values to be processed. Name of
parameter should be provided as name of each list component.}

\item{parameter_method}{Parameter designating the method to use.

Details of all_combinations method...

Details of monte-carlo method...

Details of latin hypercube method...

Details of specific values method... This option is analogous to importing
a dataframe of parameter sets, but specifies values by parameter in a list.}
}
\description{
Converts a list of parameter inputs (ranges, values, etc.) to a data frame
containing the specified number of unique parameter set combinations.
}
