% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhessys_setup.R
\name{make_rhessys_folders}
\alias{make_rhessys_folders}
\title{Functions for producing RHESSys folders}
\usage{
make_rhessys_folders(extra_folders = NA)
}
\description{
Description. Note that many RHESSys files are produced via GRASS
}
