% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tec_file.R
\name{make_tec_file}
\alias{make_tec_file}
\title{Make Tec File from R}
\usage{
make_tec_file(tec_file, tec_data)
}
\arguments{
\item{tec_file}{Name and directory of tec file to be created}

\item{tec_data}{Data frame containing the data needed for the tec file. Data
frame columns include 'year', 'month', 'day', 'hour' and the name of tec
command (e.g. "print_daily_on").}
}
\description{
Description...
}
