% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_output_filter.R
\name{modify_output_filter}
\alias{modify_output_filter}
\title{modify_output_filter}
\usage{
modify_output_filter(
  filter_in,
  timestep = NULL,
  output_format = NULL,
  output_path = NULL,
  output_filename = NULL,
  spatial_level = NULL,
  spatial_ID = NULL,
  variables = NULL
)
}
\arguments{
\item{filter_in}{Path to the yaml filter file}

\item{timestep}{The timestep of the filter: 'daily', 'monthly', 'annual', 'hourly'}

\item{output_format}{The format for the RHESSys output files to be written: 'csv' or 'netcdf'}

\item{output_path}{The path where RHESSys output will be written.}

\item{output_filename}{The filename for RHESSys output}

\item{spatial_level}{The spatial level for output to be collected at/aggregated to: 'basin', 'hillslope', 'zone', 'patch', 'stratum'}

\item{spatial_ID}{IDs to subset the spatial units by.}

\item{variables}{The RHESSys internal variable names to output. If aggregating from a finer spatial level use syntax: '<spatial>.<varaible>'}
}
\description{
Modify a single existing output filter, either from file or based on an existing R object. Will error if given multiple filters.
This could be done via normal list modification in R as well.
}
\author{
Will Burke
}
