% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_control.R
\name{output_control}
\alias{output_control}
\title{output_control}
\usage{
output_control(
  output_method,
  output_variables,
  return_data,
  output_folder,
  output_filename,
  runID = 1
)
}
\arguments{
\item{output_method}{Output method (function) to use. "awk" will use awk, "r" will use new R based method,
any other non NULL input will use the older R based output selection.}

\item{output_variables}{Datafrane with two named columns: "variables" containing variables of interest
found in the header of standard rhessys output files, "out_file" points to the files containing the associated variables,
this can be either the path to that file, can use the abbreviation for the different output types
(patch daily = pd, patch daily grow = pdg etc.) and the files in output_folder will be parsed appropriately,
or can use the fully written out space.time suffix, e.g. "patch.daily" or "grow_patch.daily"}

\item{return_data}{TRUE/FALSE if the function should return a data.table of the selected output - for now only works if doing 1 run}

\item{output_folder}{Folder where rhessys output is located}

\item{output_filename}{Base file name of standard rhessys output}

\item{runID}{Integer ID of the current run, used internally}
}
\description{
Passes output variables to appropriate output selection method.
}
