% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_grid.R
\name{parameters_grid}
\alias{parameters_grid}
\title{Create parameter sets from grid search}
\usage{
parameters_grid(m, k, m_v, k_v, pa, po, gw1, gw2, ...)
}
\description{
Create parameter sets from grid search
}

