% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_mc.R
\name{parameters_mc}
\alias{parameters_mc}
\title{Combines Monte Carlo parameter set}
\usage{
parameters_mc(m, k, m_v, k_v, pa, po, gw1, gw2, ...)
}
\description{
Combines vectors of parameter inputs into data frame of parameter sets that
\code{run_rhessys()} cycles through.
}
\details{
Each input parameter vector should be of equal length. Extra parameters beyond the
standard RHESSys calibrated parameters (e.g. def parameters) will need to be
accompanied by extra filenames for awk command. See \code{run_rhessys()} inputs.
}

