% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_fam_agg.R
\name{patch_fam_agg}
\alias{patch_fam_agg}
\title{patch_fam_agg}
\usage{
patch_fam_agg(X, areas = NULL, na.rm = FALSE)
}
\arguments{
\item{X}{Data Frame or Data Table to aggregate}

\item{areas}{Data frame (or table) of patch IDs and associated areas.
Needed for stratum outputs which don't include area.}

\item{na.rm}{Should NAs ne removed when calculating weighed average (passed through to weighted.mean)}
}
\description{
Aggregate (aspatial) patches back to spatial patches via weighted mean.
Add in exceptions if you want to aggregate via different methods.
}
