% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output_filter.R
\name{read_output_filter}
\alias{read_output_filter}
\title{read_output_filter}
\usage{
read_output_filter(filter_in)
}
\arguments{
\item{filter_in}{Path to the yaml filter file}
}
\description{
Reads a yaml format RHESSys output filter and creates a R list. Handles tabs in input file,
}
\author{
Will Burke
}
