% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rhessys_met.R
\name{read_rhessys_met}
\alias{read_rhessys_met}
\title{Imports RHESSys Climate Data}
\usage{
read_rhessys_met(prename, dates = FALSE)
}
\arguments{
\item{prename}{String giving the prefix for the climate data to be imported (e.g
"clim/seqname").}

\item{dates}{TRUE/FALSE to return just the start and end dates, to use with run_rhessys}
}
\description{
This function imports RHESSys climate data into R
}
