% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_rhessys_output.R
\name{readin_rhessys_output}
\alias{readin_rhessys_output}
\title{readin_rhessys_output}
\usage{
readin_rhessys_output(pre, read_space = "dynamic",
  read_time = "dynamic", read_grow = "dynamic", stat = FALSE,
  stat_time = c("wy", "wyd"), stat_space = "b")
}
\arguments{
\item{pre}{file prefix. Does not include spatial or temporal level of aggrigation
ie. "my_watershed" NOT "my_watershed_basin" or "my_watershed_basin.daily"}

\item{read_space}{Spatial levels of all_files to be read. Default is "dynamic" which will get all all_files
that were output (with more than 1 line) with the matching prefix. Can also input a character
vector with any combination of "b", "h", "z", "p", "c", "f". Even if all_files are specified,
they will be checked if they have data, and only read if they do.}

\item{read_time}{Timesteps of all_files to be read. Default is "dynamic". Can specify any combination of
"h", "d", "m", "y", for hourly, daily, monthly, and yearly respectively}

\item{read_grow}{Should grow files be read? Default is dynamic. Can also specify TRUE or FALSE}

\item{stat}{Stat to be computed. Only tested with "mean".}

\item{stat_time}{Not used yet - only produces "wy", "wyd" for now.}

\item{stat_space}{Not used yet. All stats aggregate to basin level. In the future will have options here
to retain finer spatial resolutions.}
}
\description{
Pulls output from a single RHESSys run into R, optionally calculates statistics. This should
be much faster than readin_rhessys_output since it uses data.table file read and aggregation methods
}
