% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_rhessys_output2.R
\name{readin_rhessys_output2}
\alias{readin_rhessys_output2}
\title{Extracts output from a multiple RHESSys runs into R}
\usage{
readin_rhessys_output2(var_names, path, initial_date, parameter_file = NULL)
}
\arguments{
\item{var_names}{Vector of the variables to be imported into R}

\item{path}{Path to the directory containing data}

\item{initial_date}{Initial date for the data e.g. ymd("1941-10-1")}

\item{z}{Optional parameters for including in analysis}
}
\description{
Imports multiple-run RHESSys output that has been consolidated by variable. This
function imports and manipulates data so that it is 'tidy' and easily imported
into ggplot.
}

