% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_rhessys_output_cal.R
\name{readin_rhessys_output_cal}
\alias{readin_rhessys_output_cal}
\title{Extracts output from a multiple RHESSys runs into R}
\usage{
readin_rhessys_output_cal(
  var_names,
  path,
  initial_date,
  timestep = "daily",
  parameter_file = NULL,
  num_layers = 1
)
}
\arguments{
\item{var_names}{Vector of the variables names that are to be imported into R. Variables should all have the same number of layers.}

\item{path}{Path to the directory containing data}

\item{initial_date}{Initial date for the data e.g. lubridate::ymd("1941-10-01")}

\item{timestep}{Timestep used for modeling (e.g. yearly, monthly, or daily). Default is daily.}

\item{parameter_file}{Optional file containing parameters to be included in analysis (e.g. RHESSysIOinR output x_parameter_sets.csv)}

\item{num_layers}{Number of layers in data. For most output (e.g. patch, basin), this will generally have a value of one. The exception being when using two canopies.}
}
\description{
Imports multiple-run RHESSys output that has been consolidated by variable. This
function imports and manipulates data so that it is 'tidy' and easily imported
into ggplot.
}
\details{
This function processes the output from multiple run calibration or simulations which consolidate a selection of the output into the allsim folder.
}
