% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_rhessys_output_old.R
\name{readin_rhessys_output_old}
\alias{readin_rhessys_output_old}
\title{readin_rhessys_output}
\usage{
readin_rhessys_output_old(pre, b = 1, c = 0, z = 0, p = 0, h = 0,
  g = 0, f = 0, wy = 1, dyn_read = FALSE, clean = TRUE)
}
\arguments{
\item{pre}{file prefix. Does not include spatial or temporal level of aggrigation
ie. "my_watershed" NOT "my_watershed_basin" or "my_watershed_basin.daily"}

\item{b}{basin}

\item{c}{canopy}

\item{z}{basin}

\item{p}{canopy}

\item{h}{basin}

\item{g}{canopy}

\item{f}{fire}

\item{wy}{wateryear}

\item{dyn_read}{dynamically read all the output that exists}
}
\description{
Pulls output from a single RHESSys run into R
}
