% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys2.R
\name{run_rhessys2}
\alias{run_rhessys2}
\title{Run RHESSys model simulations}
\usage{
run_rhessys2(
  input_rhessys,
  hdr_files,
  def_pars,
  std_pars,
  tec_data,
  output_method = NULL,
  output_variables = NULL,
  return_data = FALSE
)
}
\arguments{
\item{input_rhessys}{List containing the following named elements: "rhessys_version" (path to rhessys binary),
"tec_file"(name for tec file to be built), "world_file"(path to existing worldfile), "world_hdr_prefix"(prefix for headers to create),
"flow_file"(path to existing flowtable), "start_date"(format c('yyyy mm dd hr')), "end_date"(format c('yyyy mm dd hr')),
"output_folder"(path to output folder), "output_filename"(prefix for output files to create), "command_options"(additional commandline options)}

\item{hdr_files}{List of named elements, named for each def file type (basin_def, hillslope_def, zone_def, soil_def, landuse_def, patch_def,
stratum_def) as well as an element named "base_stations". Each element should contain the path to the corresponding def file.}

\item{def_pars}{To overwrite def file parameters. Format is a list of lists, with each sub-list having the format:
list(<path to def/just use the input_hdr_list$yourheader>, <parameter name>, <value>)}

\item{std_pars}{List of standard (command line) parameters}

\item{tec_data}{Input tec events, see input_tec function}

\item{output_method}{"awk" will use awk, any other non NULL input will use R based output selection}

\item{output_variables}{List of output variables to subset via the indicated method, Defaults to NULL}

\item{return_data}{TRUE/FALSE if the function should return a data.table of the selected output - for now only works if doing 1 run}
}
\description{
Simple version of run_rhessys. No sensitivity for now.
}
