% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_output_variables_R.R
\name{select_output_variables_R}
\alias{select_output_variables_R}
\title{Selects RHESSys output variables}
\usage{
select_output_variables_R(
  output_variables,
  output_folder,
  output_filename,
  run,
  output_initiation,
  return_data = FALSE
)
}
\arguments{
\item{output_variables}{Datafrane with named columns: "variables" containing variables matching variables of interest
found in the header of rhessys output files, "out_file" points to the files containing the associated variables,
this can be either the path to that file, can use the abbreviation for the different output types (patch daily = pd, patch daily grow = pdg etc.)
and the files in output_folder will be parsed appropriately, or can use the fully written out space.time suffix, e.g. "patch.daily" or "grow_patch.daily"}

\item{output_folder}{Folder where rhessys output is located (e.g. 'out')}

\item{run}{Simulation number. Used to reset files in allsim at beginning of simulation}

\item{return_data}{TRUE/FALSE if the function should return a data.table of the selected output - for now only works if doing 1 run}
}
\description{
Somewhat optimized R-based read and subsetting of RHESSys output
}
