% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_output_variables_w_awk.R
\name{select_output_variables_w_awk}
\alias{select_output_variables_w_awk}
\title{Selects RHESSys output variables}
\usage{
select_output_variables_w_awk(output_variables, output_folder, run,
  output_initiation)
}
\arguments{
\item{output_variables}{A data frame containing containing variable of interest, location/name of awk file (relative to
output_file location), and the location/name of rhessys output file with variable
of interest.}

\item{output_folder}{Folder where rhessys output is located (e.g. 'out')}

\item{run}{Simulation number. Used to reset files in allsim at beginning of simulation}
}
\description{
This function does etc, etc... Requires that an awk file be created that indicates
the columns that correspond to the variable(s) of interest. See \code{create_awk}
}
