% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_evaluation_criteria.R
\name{select_parameter_sets}
\alias{select_parameter_sets}
\title{Select approach to evaluate objective function output}
\usage{
select_parameter_sets(
  obj_function_all,
  obj_function_selection,
  method,
  num_ps = 1,
  ...
)
}
\description{
Description...
}
