% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_tec_std.R
\docType{data}
\name{start}
\alias{start}
\title{IOin_tec_std}
\format{
An object of class \code{Date} of length 1.
}
\usage{
start
}
\arguments{
\item{start}{start date of run}

\item{end}{End date of run - the last entire day to be run}

\item{output_state}{TRUE/FALSE if an output_current_state tec event should be scheduled at the end of the simulation}
}
\description{
Input function to construct a dataframe of standard tec events, including start, grow start, and output state
}
\author{
Will Burke
}
\keyword{datasets}
