% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_parameters.R
\name{substitute_parameters}
\alias{substitute_parameters}
\title{Selects parameter (and variables) to be dynamically searched}
\usage{
substitute_parameters(par_value, awk_file, input_file, output_folder)
}
\arguments{
\item{par_value}{??}

\item{awk_file}{??}

\item{input_file}{??}

\item{output_folder}{??}
}
\description{
This function calls an awk file for substituting parameter values in RHESSys
}

