% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tec_repeat.R
\name{tec_repeat}
\alias{tec_repeat}
\title{tec_repeat}
\usage{
tec_repeat(
  start,
  end,
  interval = NULL,
  unit,
  event_name,
  world = NULL,
  redefine = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{start}{Vector containing start date. Format is "c(year,month,day,hour)".}

\item{end}{Vector containing end date. Format is "c(year,month,day,hour)".}

\item{interval}{Interval to repeat chosen tec event. If left NULL will be executed 1 time.}

\item{unit}{Unit of time for the repeat interval}

\item{event_name}{The name/type of tec event to repeat.}

\item{world}{Worldfile basename to modify and repeat}

\item{redefine}{Redefine file to copy and rename. If NULL no files copied.}

\item{overwrite}{Should existing files be overwritten}
}
\description{
Repeats tec events at chosen intervals. Outputs a dataframe of the events for use in the tec input for RHESSysIOinR,
optionally will copy an input redefine file with generated filenames.
}
\author{
Will Burke
}
