% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output_filter.R
\name{write_output_filter}
\alias{write_output_filter}
\title{write_output_filter}
\usage{
write_output_filter(output_filter, runID = NULL, return_timesteps = FALSE)
}
\arguments{
\item{output_filter}{An output filter, either an R list with 1 to n number of filters read in/modified/generated via IOin_output_filter.R
(or associated functions - build_output_filter.R, read_output_filter.R, modify_output_filter.R), or a file path pointing to an
existing output filter.}

\item{runID}{The unique ID used to track input and output files if running multiple scenarios, and thus multiple instances of run_rhessys_core.}
}
\description{
Writes an output filter file based on an input R list containing the needed elements.
}
\author{
Will Burke
}
