% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_auto.R
\name{clim_auto}
\alias{clim_auto}
\title{clim_auto}
\usage{
clim_auto(
  base_station_id,
  x_coordinate,
  y_coordinate,
  z_coordinate,
  effective_lai,
  screen_height,
  annual_prefix = "annual",
  monthly_prefix = "monthly",
  daily_prefix = "daily",
  hourly_prefix = "hourly"
)
}
\description{
Generate input for run_rhessys basestation input
}
\author{
Will Burke
}
