% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rhessys.R
\name{compile_rhessys}
\alias{compile_rhessys}
\title{compile_rhessys}
\usage{
compile_rhessys(
  location,
  delete_objs = TRUE,
  destination = NULL,
  make_args = NULL,
  CFLAGS = NULL,
  ignore.stdout = FALSE
)
}
\arguments{
\item{location}{The file patch to where the rhessys makefile or folder is}

\item{destination}{Optional input of where to move resulting RHESSys executable}

\item{CFLAGS}{Overwrite CFLAGS in the makefile}

\item{ignore.stdout}{Passed through to system()}

\item{delete_obj}{TRUE/FALSE to delete objects (before and after)}
}
\description{
Compiles rhessys, with options to delete objects and move resulting RHESSys executable
}
\author{
Will Burke
}
