% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input_files.R
\name{generate_input_files}
\alias{generate_input_files}
\title{Produces a hdr file in RHESSys}
\usage{
generate_input_files(input_rhessys, input_hdr_list, input_clim_base_list,
  input_tec_data, option_sets_all, option_sets_def_par, option_sets_par,
  option_sets_hdr, option_sets_dated_seq)
}
\description{
The hdr file in RHESSys determines which definition and climate files are to be used for a given run.
}
\details{
This function produces hdr file and associated files that are called by hdr file; def files and clim.base files.
}
