% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tec.R
\name{input_tec}
\alias{input_tec}
\title{input_tec}
\usage{
input_tec(...)
}
\arguments{
\item{...}{Vector(s) containing tec events in format c(<year>,<month>,<day>,<hour>,<name/event type>).
Additional tec events can be added just as additional arguments.}
}
\description{
Tiny wrapper to simplify input of tec events
}
\author{
Will Burke
}
