% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tec.R
\name{input_tec}
\alias{input_tec}
\title{input_tec}
\usage{
input_tec(..., start_end = NULL)
}
\arguments{
\item{...}{Vector(s) containing tec events in format c(<year>,<month>,<day>,<hour>,<name/event type>).
Additional tec events can be added just as additional arguments.}

\item{start_end}{Vector of two charaters, indicating start and end dates, same as used to input start and end date for input_rhessys.
Used to auto generate tec events for printing (normal and grow), and output current state at end}
}
\description{
Tiny wrapper to simplify input of tec events
}
\author{
Will Burke
}
