% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOin_tec_std.R
\name{IOin_tec_std}
\alias{IOin_tec_std}
\title{IOin_tec_std}
\usage{
IOin_tec_std(start_end = NULL, start = NULL, end = NULL, output_state = TRUE)
}
\arguments{
\item{start_end}{Vector of two charaters, indicating start and end dates, same as used to input start and end date for input_rhessys.
Used to auto generate tec events for printing (normal and grow), and output current state at end}

\item{start}{start date of run - used along with 'end' arg}

\item{end}{End date of run}

\item{output_state}{TRUE/FALSE if an output_current_state tec event should be scheduled at the end of the simulation}
}
\description{
Input function to construct a dataframe of standard tec events, including start, grow start, and output state
}
\author{
Will Burke
}
