% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hdr_file.R
\name{make_hdr_file}
\alias{make_hdr_file}
\title{Produces components of a hdr file}
\usage{
make_hdr_file(master_table, path_initial, num_files, default_file)
}
\arguments{
\item{master_table}{Data frame of a single row with references to def file
number. A value of zero for the def file number indicates that the original
def file is to be used. Higher numbers are added as a extension to the file
name.}

\item{path_initial}{Path to the non-differiated def file}

\item{num_files}{???}

\item{default_file}{???}
}
\description{
This function generates the lines in a hdr file associated with a single def
file.
}
