% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hdr_file.R
\name{make_hdr_file}
\alias{make_hdr_file}
\title{Write Header File}
\usage{
make_hdr_file(input_rhessys, hdr_files, def_files, runID)
}
\arguments{
\item{input_rhessys}{List containing the following named elements: "rhessys_version" (path to rhessys binary),
"tec_file"(name for tec file to be built), "world_file"(path to existing worldfile), "world_hdr_prefix"(prefix for headers to create),
"flow_file"(path to existing flowtable), "start_date"(format c('yyyy mm dd hr')), "end_date"(format c('yyyy mm dd hr')),
"output_folder"(path to output folder), "output_filename"(prefix for output files to create), "command_options"(additional commandline options)}

\item{hdr_files}{List of named elements, named for each def file type (basin_def, hillslope_def, zone_def, soil_def, landuse_def, patch_def,
stratum_def) as well as an element named "base_stations". Each element should contain the path to the corresponding def file.}

\item{def_files}{Data frame of def file parameter changes}

\item{runID}{The unique ID used to track input and output files if running multiple scenarios, and thus multiple instances of run_rhessys_core.}
}
\description{
New version of header file creation function.
}
\author{
Will Burke
}
