% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_control.R
\name{output_control}
\alias{output_control}
\title{output_control}
\usage{
output_control(
  output_method,
  output_variables,
  return_data,
  output_folder,
  output_filename,
  run = 1
)
}
\arguments{
\item{output_method}{"awk" will use awk, any other non NULL input will use R based output selection}

\item{output_variables}{List of output variables to subset via the indicated method, Defaults to NULL}

\item{return_data}{TRUE/FALSE if the function should return a data.table of the selected output - for now only works if doing 1 run}

\item{output_folder}{Folder where rhessys output is located (e.g. 'out')}

\item{output_filename}{Base file name of standard rhessys output}

\item{run}{Simulation number. Used to reset files in allsim at beginning of simulation}
}
\description{
Passes output variables to appropriate output method - eventually should put all the output code in here directly maybe?
}
