% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_rhessys_output.R
\name{readin_rhessys_output}
\alias{readin_rhessys_output}
\title{Pulls output from a single RHESSys run into R}
\usage{
readin_rhessys_output(pre, b = 1, c = 0, z = 0, p = 0, h = 0, g = 0,
  wy = 1)
}
\arguments{
\item{pre}{file}

\item{b}{basin}

\item{c}{canopy}

\item{z}{basin}

\item{p}{canopy}

\item{h}{basin}

\item{g}{canopy}

\item{wy}{wateryear}
}
\description{
It does fanciness...
}

