% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhessys_command.R
\name{rhessys_command}
\alias{rhessys_command}
\title{Executes single RHESSys run on command line}
\usage{
rhessys_command(rhessys_version, world_file, world_hdr_file, tec_file,
  flow_file, start_date, end_date, output_file, input_parameters,
  command_options)
}
\arguments{
\item{rhessys_version}{path and file of compiled version of RHESSys.}

\item{world_file}{path and file of world_file. World_file is produced using
grass-to-world}

\item{world_hdr_file}{???}

\item{tec_file}{???}

\item{flow_file}{???}

\item{start_date}{???}

\item{end_date}{???}

\item{output_file}{???}

\item{input_parameters}{???}

\item{command_options}{???}
}
\description{
\code{rhessys_command} calls rhessys on the command line.
}
\details{
Requires that supporting files be developed prior to running RHESSys. These
include ...
}
