% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhessys_command.R
\name{rhessys_command}
\alias{rhessys_command}
\title{Executes single RHESSys run on command line}
\usage{
rhessys_command(
  rhessys_version,
  world_file,
  world_hdr_file,
  tec_file,
  flow_file,
  start_date,
  end_date,
  output_file = NULL,
  input_parameters,
  output_filter = NULL,
  command_options,
  prefix_command = NULL,
  return_cmd = FALSE
)
}
\arguments{
\item{rhessys_version}{Path and file name of compiled version of RHESSys.}

\item{world_file}{Oath and file name of RHESSys world_file.}

\item{world_hdr_file}{Path and file name of RHESSys header file}

\item{tec_file}{Path and file name of RHESSys temporal event control (tec) file}

\item{flow_file}{Path and file name of RHESSys flow table file}

\item{start_date}{Start date character vector in format <year> <month> <day> <hour>, delimited by spaces. Ex. '1990 12 30 01'}

\item{end_date}{End date character vector, same format as start_date}

\item{output_file}{Path and base file name of RHESSys output}

\item{input_parameters}{Soil parameters passed to RHESSys command line.}

\item{output_filter}{Path to a yaml formatted output filter.}

\item{command_options}{RHESSys command line options, ex. '-g' or '-p'.}

\item{prefix_command}{A shell command to be run previous to the RHESSys command line call.}

\item{return_cmd}{true/false passed from run_rhessys_single
This can be used to source a shell script, which itself can run multiple commands if needed.}
}
\description{
\code{rhessys_command} Assembles command line RHESSys call, and runs it. See the RHESSys wiki:
https://github.com/RHESSys/RHESSys/wiki/RHESSys-command-line-options
}
