% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHESSysIOinR.R
\docType{package}
\name{RHESSysIOinR}
\alias{-package}
\alias{RHESSysIOinR}
\alias{.datatable.aware}
\title{RHESSysIOinR}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
RHESSysIOinR contains functions for running [RHESSys](https://github.com/RHESSys/RHESSys) in R
and processing output. The objective of this package is to clearly and efficiently produce R code
that can be used to setup RHESSys, conduct calibrations and simulations, and process output. This
package supports a long(er)-term goal of having experiments involving RHESSys be reproducible.
}
\keyword{datasets}
