% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys.R
\name{run_rhessys}
\alias{run_rhessys}
\title{Run RHESSys model simulations}
\usage{
run_rhessys(
  parameter_method,
  output_method,
  input_rhessys,
  input_hdr_list,
  input_preexisting_table = NULL,
  input_def_list,
  input_standard_par_list,
  input_clim_base_list,
  input_dated_seq_list = NULL,
  input_tec_data,
  output_variables = NULL,
  output_initiation = 1,
  return_data = FALSE
)
}
\arguments{
\item{parameter_method}{One of the following methods: "all_combinations","lhc","monte_carlo","exact_values"}

\item{output_method}{"awk" will use awk, any other non NULL input will use R based output selection}

\item{input_rhessys}{List containing the following named elements: "rhessys_version" (path to rhessys binary),
"tec_file"(name for tec file to be built), "world_file"(path to existing worldfile), "world_hdr_prefix"(prefix for headers to create),
"flow_file"(path to existing flowtable), "start_date"(format c('yyyy mm dd hr')), "end_date"(format c('yyyy mm dd hr')),
"output_folder"(path to output folder), "output_filename"(prefix for output files to create), "command_options"(additional commandline options)}

\item{input_hdr_list}{List of named elements, named for each def file type (basin_def, hillslope_def, zone_def, soil_def, landuse_def, patch_def,
stratum_def) as well as an element named "base_stations". Each element should contain the path to the corresponding def file.}

\item{input_preexisting_table}{Optional means of inputing parameters based on a prexisting setup. Defaults to NULL}

\item{input_def_list}{To overwrite def file parameters. Format is a list of lists, with each sub-list having the format:
list(<path to def/just use the input_hdr_list$yourheader>, <parameter name>, <value>)}

\item{input_standard_par_list}{List of standard (command line) parameters}

\item{input_clim_base_list}{List of the climate base station components - also see clim_auto}

\item{input_dated_seq_list}{List of dated sequences, defaults to NULL}

\item{input_tec_data}{Input tec events, see input_tec function}

\item{output_variables}{List of output variables to subset via the indicated method, Defaults to NULL}

\item{return_data}{TRUE/FALSE if the function should return a data.table of the selected output - for now only works if doing 1 run}
}
\description{
Permits the calibration and simulation of the rhessys
model. Multiple model runs can be run, either in series or in parallel (such
as with cluster computing).
}
