% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys.R
\name{run_rhessys}
\alias{run_rhessys}
\title{Run multiple RHESSys model simulations}
\usage{
run_rhessys(parameter_method, input_rhessys, input_hdr_list,
  input_preexisting_table = NULL, input_def_list, input_standard_par_list,
  input_clim_base_list, input_dated_seq_list = NULL, input_tec_data,
  output_variables = NULL, output_initiation = 1)
}
\description{
\code{run_rhessys} permits the calibration and simulation of the rhessys
model. Multiple model runs can be run, either in series or in parallel (such
as with cluster computing).
}
