% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys.R
\name{run_rhessys}
\alias{run_rhessys}
\title{Run multiple RHESSys model simulations}
\usage{
run_rhessys(parameter_method = c("all_combinations", "lhc", "monte_carlo",
  "exact_values"), input_rhessys, input_hdr_list, input_preexisting_table,
  input_def_list, input_standard_par_list, input_clim_base_list,
  input_dated_seq_list, input_tec_data, output_variables)
}
\description{
\code{run_rhessys} permits the calibration and simulation of the rhessys
model. Multiple model runs can be run, either in series or in parallel (such
as with cluster computing).
}
