% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys_multi.R
\name{run_rhessys_multi}
\alias{run_rhessys_multi}
\title{Run multiple RHESSys simulations}
\usage{
run_rhessys_multi(
  input_rhessys,
  hdr_files,
  tec_data,
  def_pars = NULL,
  clim_base = NULL,
  output_filter = NULL,
  return_cmd = FALSE,
  return_file_list = FALSE,
  par_option = TRUE,
  runIDoffset = 0,
  parallel = TRUE,
  n_cores = NULL,
  write_log = FALSE,
  log_loc = "~/rhessys_run_log.csv"
)
}
\arguments{
\item{input_rhessys}{List containing the following named elements: "rhessys_version" (path to rhessys binary),
"tec_file"(name for tec file to be built), "world_file"(path to existing worldfile), "world_hdr_prefix"(prefix for headers to create),
"flow_file"(path to existing flowtable), "start_date"(format c('yyyy mm dd hr')), "end_date"(format c('yyyy mm dd hr')),
"output_folder"(path to output folder), "output_filename"(prefix for output files to create), "command_options"(additional commandline options)}

\item{hdr_files}{List of named elements, named for each def file type (basin_def, hillslope_def, zone_def, soil_def, landuse_def, patch_def,
stratum_def) as well as an element named "base_stations". Each element should contain the path to the corresponding def file.}

\item{tec_data}{Input tec events, see input_tec function}

\item{def_pars}{To overwrite def file parameters. Format is a list of lists, with each sub-list having the format:
list(<path to def/just use the input_hdr_list$yourheader>, <parameter name>, <value>). Defaults to NULL}

\item{clim_base}{Data for input climate basestation to be written. Defaults to NULL, which assumes you havean existing basestation pointed
to in input_rhessys.}

\item{output_filter}{An output filter, either an R list with 1 to n number of filters read in/modified/generated via IOin_output_filter.R
(or associated functions - build_output_filter.R, read_output_filter.R, modify_output_filter.R), or a file path pointing to an
existing output filter.}

\item{return_cmd}{TRUE/FALSE if run_rhessys_single should return the command line call INSTEAD of running.}

\item{return_file_list}{TRUE/FALSE if run_rhessys_single should return a list of output files generated.}

\item{par_option}{TRUE/FALSE if the -par command line options should be used if running multiple runs and outputting current worldfile state.
Can also just set as a number to set specific par ID.}

\item{runIDoffset}{An integer to offset the run IDs. Useful for continuing runs from a previous batch.}

\item{parallel}{Defaults to TRUE. Should the parallel package be used to parallelize the rhessys runs.}

\item{n_cores}{The number of cores to use in a parallelized cluster. If left NULL, will autodetect number of cores and use total - 1.}

\item{write_log}{TRUE/FALSE Writes or appends a log to a specified file, by row}
}
\description{
Runs RHESSys simulations.
}
\author{
Will Burke
}
