% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys_multi.R
\name{run_rhessys_multi}
\alias{run_rhessys_multi}
\title{Run multiple RHESSys simulations}
\usage{
run_rhessys_multi(
  input_rhessys,
  hdr_files,
  tec_data,
  def_pars = NULL,
  output_filter = NULL,
  runID = NULL,
  parallel = TRUE,
  n_cores = 2
)
}
\arguments{
\item{input_rhessys}{List containing the following named elements: "rhessys_version" (path to rhessys binary),
"tec_file"(name for tec file to be built), "world_file"(path to existing worldfile), "world_hdr_prefix"(prefix for headers to create),
"flow_file"(path to existing flowtable), "start_date"(format c('yyyy mm dd hr')), "end_date"(format c('yyyy mm dd hr')),
"output_folder"(path to output folder), "output_filename"(prefix for output files to create), "command_options"(additional commandline options)}

\item{hdr_files}{List of named elements, named for each def file type (basin_def, hillslope_def, zone_def, soil_def, landuse_def, patch_def,
stratum_def) as well as an element named "base_stations". Each element should contain the path to the corresponding def file.}

\item{tec_data}{Input tec events, see input_tec function}

\item{def_pars}{To overwrite def file parameters. Format is a list of lists, with each sub-list having the format:
list(<path to def/just use the input_hdr_list$yourheader>, <parameter name>, <value>). Defaults to NULL}

\item{output_filter}{An output filter, either an R list with 1 to n number of filters read in/modified/generated via IOin_output_filter.R
(or associated functions - build_output_filter.R, read_output_filter.R, modify_output_filter.R), or a file path pointing to an
existing output filter.}

\item{runID}{The unique ID used to track input and output files if running multiple scenarios, and thus multiple instances of run_rhessys_core.}

\item{parallel}{Defaults to TRUE. Should the parallel package be used to parallelize the rhessys runs.}

\item{n_cores}{The number of cores to use in a parallelized cluster.}
}
\description{
Runs RHESSys simulations.
}
\author{
Will Burke
}
