% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys_multi.R
\name{run_rhessys_multi}
\alias{run_rhessys_multi}
\title{Run multiple RHESSys simulations}
\usage{
run_rhessys_multi(
  input_rhessys,
  hdr_files,
  std_pars,
  tec_data,
  def_pars = NULL,
  clim_base = NULL,
  output_method = NULL,
  output_variables = NULL,
  return_data = FALSE,
  runID = NULL
)
}
\arguments{
\item{input_rhessys}{List containing the following named elements: "rhessys_version" (path to rhessys binary),
"tec_file"(name for tec file to be built), "world_file"(path to existing worldfile), "world_hdr_prefix"(prefix for headers to create),
"flow_file"(path to existing flowtable), "start_date"(format c('yyyy mm dd hr')), "end_date"(format c('yyyy mm dd hr')),
"output_folder"(path to output folder), "output_filename"(prefix for output files to create), "command_options"(additional commandline options)}

\item{hdr_files}{List of named elements, named for each def file type (basin_def, hillslope_def, zone_def, soil_def, landuse_def, patch_def,
stratum_def) as well as an element named "base_stations". Each element should contain the path to the corresponding def file.}

\item{std_pars}{List of standard (command line) parameters.}

\item{tec_data}{Input tec events, see input_tec function}

\item{def_pars}{To overwrite def file parameters. Format is a list of lists, with each sub-list having the format:
list(<path to def/just use the input_hdr_list$yourheader>, <parameter name>, <value>). Defaults to NULL}

\item{clim_base}{Data for input climate basestation to be written. Defaults to NULL, which assumes you havean existing basestation pointed
to in input_rhessys.}

\item{output_method}{"awk" will use awk, any other non NULL input will use R based output selection}

\item{output_variables}{List of output variables to subset via the indicated method, Defaults to NULL}

\item{return_data}{TRUE/FALSE if the function should return a data.table of the selected output - for now only works if doing 1 run}

\item{runID}{The unique ID used to track input and output files if running multiple scenarios, and thus multiple instances of run_rhessys_core.}
}
\description{
Runs RHESSys simulations.
}
\author{
Will Burke
}
