% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rhessys_single.R
\name{run_rhessys_single}
\alias{run_rhessys_single}
\title{Run RHESSys simulation}
\usage{
run_rhessys_single(
  input_rhessys,
  hdr_files,
  tec_data = NULL,
  def_pars = NULL,
  clim_base = NULL,
  output_filter = NULL,
  cmd_pars = NULL,
  par_option = TRUE,
  return_cmd = FALSE,
  return_file_list = FALSE,
  docker_image = NULL,
  docker_host_dir = NULL,
  docker_container_dir = NULL,
  docker_run_options = NULL,
  write_run_metadata = FALSE,
  write_log = FALSE,
  log_loc = "~/rhessys_run_log.csv",
  runID = NULL
)
}
\arguments{
\item{input_rhessys}{List containing the following named elements: "rhessys_version" (path to rhessys binary),
"tec_file"(name for tec file to be built), "world_file"(path to existing worldfile), "world_hdr_prefix"(prefix for headers to create),
"flow_file"(path to existing flowtable), "start_date"(format c('yyyy mm dd hr')), "end_date"(format c('yyyy mm dd hr')),
"output_folder"(path to output folder), "output_filename"(prefix for output files to create), "command_options"(additional commandline options)}

\item{hdr_files}{List of named elements, named for each def file type (basin_def, hillslope_def, zone_def, soil_def, landuse_def, patch_def,
stratum_def) as well as an element named "base_stations". Each element should contain the path to the corresponding def file.}

\item{tec_data}{Input tec events, see input_tec function}

\item{def_pars}{To overwrite def file parameters. Format is a list of lists, with each sub-list having the format:
list(<path to def/just use the input_hdr_list$yourheader>, <parameter name>, <value>). Defaults to NULL}

\item{clim_base}{Data for input climate basestation to be written. Defaults to NULL, which assumes you havean existing basestation pointed
to in input_rhessys.}

\item{output_filter}{An output filter, either an R list with 1 to n number of filters read in/modified/generated via IOin_output_filter.R
(or associated functions - build_output_filter.R, read_output_filter.R, modify_output_filter.R), or a file path pointing to an
existing output filter.}

\item{cmd_pars}{List of standard (command line) parameters.}

\item{par_option}{TRUE/FALSE if the -par command line options should be used if running multiple runs and outputting current worldfile state.
Can also just set as a number to set specific par ID.}

\item{return_cmd}{TRUE/FALSE if run_rhessys_single should return the command line call INSTEAD of running.}

\item{return_file_list}{TRUE/FALSE if run_rhessys_single should return a list of output files generated.}

\item{docker_image}{Name (and optionally tag) of a docker image containing RHESSys, e.g. \code{"rhessys:develop"}. If supplied, a fresh, disposable container is started for this single run via \code{docker run --rm} (no \code{--name} is set, so many calls can be launched in parallel without colliding). \code{docker_host_dir} is bind-mounted into the container so the relative file paths (world_file, tec_file, etc.) resolve correctly, and the container is removed automatically when the run finishes.}

\item{docker_host_dir}{Path on the host machine to bind-mount into the container - this should (probably) be your project directory, the directory the relative file paths (world_file, tec_file, etc.) are written relative to. Defaults to the current working directory (\code{getwd()}) if NULL. Converted to an absolute path internally, since docker requires this for bind mounts.}

\item{docker_container_dir}{Path inside the container to mount \code{docker_host_dir} to, and the working directory the RHESSys call is run from inside the container. Defaults to \code{"/mnt/rhessys_run"} if NULL.}

\item{docker_run_options}{Optional additional flags appended to \code{docker run}, e.g. \code{"--cpus=2 --memory=4g"} to cap resources for a given parallel run.}

\item{write_run_metadata}{TRUE/FALSE if a text file containing run metadata should be written to the same location as your output.}

\item{write_log}{TRUE/FALSE Writes or appends a log to a specified file, by row}

\item{runID}{The unique ID used to track input and output files if running multiple scenarios, and thus multiple instances of run_rhessys_core.}

\item{log_log}{Location of the log to be written/appended}
}
\description{
Writes files and assembles inputs to run a single instance of RHESSys.
}
