% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_canopy_output.R
\name{separate_canopy_output}
\alias{separate_canopy_output}
\title{Separate Canopy Output by Canopy}
\usage{
separate_canopy_output(variable, dates, num_canopies = 2)
}
\arguments{
\item{variable}{Variable of interest from canopy output.}

\item{dates}{Date column from canopy output}

\item{num_canopies}{Number of canopy (stratum) per patch}
}
\description{
Function separates RHESSys canopy output by canopy and outputs a data frame that
may be analyzed via \code{ggplot}.
}

