% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tec_repeat.R
\name{tec_repeat}
\alias{tec_repeat}
\title{tec_repeat}
\usage{
tec_repeat(start, end, repeat_interval, tec_type, world, thin_mult = NULL,
  thin_type = NULL, random = NULL)
}
\arguments{
\item{start}{Vector containing start date. Format is "c(year,month,day,hour)".}

\item{end}{Vector containing end date. Format is "c(year,month,day,hour)".}

\item{repeat_interval}{Interval in years to repeat chosen tec event.}

\item{tec_type}{Type of tec event to scheduled.}

\item{world}{Worldfile.}

\item{thin_mult}{Multiplier to use for thinning.}

\item{thin_type}{Type of thinning event.}

\item{random}{Set as "single" to use a single random date for all events, "all" to randomize all dates.}
}
\description{
Repeats tec events at chosen intervals.
}
\author{
Will Burke
}
