% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tec_repeat.R
\name{tec_repeat}
\alias{tec_repeat}
\title{tec_repeat}
\usage{
tec_repeat(start, end, repeat_interval, world, mult = NULL,
  thin_type = NULL, random = NULL, write = FALSE)
}
\arguments{
\item{start}{Vector containing start date. Format is "c(year,month,day,hour)".}

\item{end}{Vector containing end date. Format is "c(year,month,day,hour)".}

\item{repeat_interval}{Interval in years to repeat chosen tec event.}

\item{world}{Worldfile.}

\item{mult}{Multiplier to use for thinning.}

\item{thin_type}{Type of thinning event.}

\item{random}{Set as "single" to use a single random date for all events, "all" to randomize all dates.}

\item{write}{TRUE/FALSE write deta doc of output to file}

\item{tec_type}{Type of tec event to scheduled.}
}
\description{
Repeats tec events at chosen intervals. Template should use -1111 for replacement value, and use worldfile name followed by "template"
}
\author{
Will Burke
}
